+++ swuix/headers/swuix/widgets/container.hpp
@@ -29,7 +29,7 @@ public:
 		return children[i];
 	}
 
-	void render(Window *window, int off_x, int off_y) {
+	void render(Window *window, float off_x, float off_y) {
 		const int n = (int)child_count();
 		for (int i = n - 1; i >= 0; --i) {
 			Widget *ch = child_at(i);
